move store from config to the router level
